refactor: Drop `MutatorSetKernel`
Pesky. It does not make sense for archival and accumulator variants to
share the same logic. The archival variant should be concerned with
storing the additional data.

Useful functions and tests have been moved to `mutator_set.rs`.
