+++ src/models/blockchain/transaction/validity/tasm/compute_canonical_commitment.rs
@@ -17,9 +17,7 @@ use tasm_lib::traits::function::{Function, FunctionInitialState};
 use triton_vm::prelude::{triton_asm, BFieldElement, Digest};
 use twenty_first::shared_math::bfield_codec::BFieldCodec;
 
-use crate::util_types::mutator_set::{
-    ms_membership_proof::MsMembershipProof, mutator_set_scheme::commit,
-};
+use crate::util_types::mutator_set::{commit, ms_membership_proof::MsMembershipProof};
 
 /// Compute a canonical commitment from an item and its membership proof.
 #[derive(Debug, Clone)]
