+++ src/util_types/mutator_set/addition_record.rs
@@ -33,12 +33,11 @@ pub fn pseudorandom_addition_record(seed: [u8; 32]) -> AdditionRecord {
 
 #[cfg(test)]
 mod addition_record_tests {
-    use crate::{
-        models::blockchain::shared::Hash, util_types::mutator_set::mutator_set_scheme::commit,
-    };
 
     use rand::random;
     use twenty_first::util_types::algebraic_hasher::AlgebraicHasher;
+
+    use crate::{models::blockchain::shared::Hash, util_types::mutator_set::commit};
 
     use super::*;
