+++ src/util_types/mutator_set/removal_record.rs
@@ -20,7 +20,6 @@ use twenty_first::util_types::algebraic_hasher::AlgebraicHasher;
 
 use super::chunk_dictionary::{pseudorandom_chunk_dictionary, ChunkDictionary};
 use super::mutator_set_accumulator::MutatorSetAccumulator;
-use super::mutator_set_kernel::MutatorSetKernel;
 use super::shared::{
     get_batch_mutation_argument_for_removal_record, indices_to_hash_map, BATCH_SIZE, CHUNK_SIZE,
     NUM_TRIALS,
@@ -144,15 +143,15 @@ impl RemovalRecord {
         removal_records: &mut [&mut Self],
         mutator_set: &MutatorSetAccumulator,
     ) {
-        let new_item_index = mutator_set.kernel.aocl.count_leaves();
+        let new_item_index = mutator_set.aocl.count_leaves();
 
         // if window does not slide, do nothing
-        if !MutatorSetKernel::window_slides(new_item_index) {
+        if !MutatorSetAccumulator::window_slides(new_item_index) {
             return;
         }
 
         // window does slide
-        let new_chunk = mutator_set.kernel.swbf_active.slid_chunk();
+        let new_chunk = mutator_set.swbf_active.slid_chunk();
         let new_chunk_digest: Digest = Hash::hash(&new_chunk);
 
         // Insert the new chunk digest into the accumulator-version of the
@@ -161,7 +160,7 @@ impl RemovalRecord {
         // a whole archival MMR for this operation, as the archival MMR can be in the
         // size of gigabytes, whereas the MMR accumulator should be in the size of
         // kilobytes.
-        let mut mmra: MmrAccumulator<Hash> = mutator_set.kernel.swbf_inactive.to_accumulator();
+        let mut mmra: MmrAccumulator<Hash> = mutator_set.swbf_inactive.to_accumulator();
         let new_swbf_auth_path: mmr::mmr_membership_proof::MmrMembershipProof<Hash> =
             mmra.append(new_chunk_digest);
 
@@ -241,9 +240,9 @@ impl RemovalRecord {
         // Perform the update of all the MMR membership proofs contained in the removal records
         mmr::mmr_membership_proof::MmrMembershipProof::<Hash>::batch_update_from_append(
             &mut mmr_membership_proofs_for_append,
-            mutator_set.kernel.swbf_inactive.count_leaves(),
+            mutator_set.swbf_inactive.count_leaves(),
             new_chunk_digest,
-            &mutator_set.kernel.swbf_inactive.get_peaks(),
+            &mutator_set.swbf_inactive.get_peaks(),
         );
     }
 
@@ -279,7 +278,7 @@ impl RemovalRecord {
     }
 
     /// Validates that a removal record is synchronized against the inactive part of the SWBF
-    pub fn validate(&self, mutator_set: &MutatorSetKernel) -> bool {
+    pub fn validate(&self, mutator_set: &MutatorSetAccumulator) -> bool {
         let peaks = mutator_set.swbf_inactive.get_peaks();
         self.target_chunks
             .dictionary
@@ -324,16 +323,16 @@ mod removal_record_tests {
     use rand::{thread_rng, Rng, RngCore};
 
     use crate::util_types::mutator_set::addition_record::AdditionRecord;
+    use crate::util_types::mutator_set::commit;
     use crate::util_types::mutator_set::ms_membership_proof::MsMembershipProof;
     use crate::util_types::mutator_set::mutator_set_accumulator::MutatorSetAccumulator;
-    use crate::util_types::mutator_set::mutator_set_scheme::commit;
     use crate::util_types::mutator_set::shared::{CHUNK_SIZE, NUM_TRIALS};
     use crate::util_types::test_shared::mutator_set::*;
 
     use super::*;
 
     fn get_item_mp_and_removal_record() -> (Digest, MsMembershipProof, RemovalRecord) {
-        let mut accumulator: MutatorSetAccumulator = MutatorSetAccumulator::default();
+        let accumulator: MutatorSetAccumulator = MutatorSetAccumulator::default();
         let (item, sender_randomness, receiver_preimage) = make_item_and_randomnesses();
         let mp: MsMembershipProof = accumulator.prove(item, sender_randomness, receiver_preimage);
         let removal_record: RemovalRecord = accumulator.drop(item, &mp);
@@ -492,12 +491,12 @@ mod removal_record_tests {
 
                 for removal_record in removal_records.iter().map(|x| &x.1) {
                     assert!(
-                        removal_record.validate(&accumulator.kernel),
+                        removal_record.validate(&accumulator),
                         "removal records must validate, i = {}",
                         i
                     );
                     assert!(
-                        accumulator.kernel.can_remove(removal_record),
+                        accumulator.can_remove(removal_record),
                         "removal records must return true on `can_remove`, i = {}",
                         i
                     );
@@ -517,18 +516,18 @@ mod removal_record_tests {
                 removal_records.choose(&mut rand::thread_rng()).unwrap();
             assert!(accumulator.verify(items[*chosen_index], &mps[*chosen_index]));
             assert!(
-                accumulator.kernel.can_remove(random_removal_record),
+                accumulator.can_remove(random_removal_record),
                 "removal records must return true on `can_remove`",
             );
             assert!(
-                random_removal_record.validate(&accumulator.kernel),
+                random_removal_record.validate(&accumulator),
                 "removal record must have valid MMR MPs"
             );
             accumulator.remove(random_removal_record);
             assert!(!accumulator.verify(items[*chosen_index], &mps[*chosen_index]));
 
             assert!(
-                !accumulator.kernel.can_remove(random_removal_record),
+                !accumulator.can_remove(random_removal_record),
                 "removal records must return false on `can_remove` after removal",
             );
         }
@@ -576,12 +575,12 @@ mod removal_record_tests {
 
             for removal_record in removal_records.iter().map(|x| &x.1) {
                 assert!(
-                    removal_record.validate(&accumulator.kernel),
+                    removal_record.validate(&accumulator),
                     "removal records must validate, i = {}",
                     i
                 );
                 assert!(
-                    accumulator.kernel.can_remove(removal_record),
+                    accumulator.can_remove(removal_record),
                     "removal records must return true on `can_remove`, i = {}",
                     i
                 );
@@ -611,11 +610,11 @@ mod removal_record_tests {
 
             for removal_record in removal_records.iter().map(|x| &x.1) {
                 assert!(
-                    removal_record.validate(&accumulator.kernel),
+                    removal_record.validate(&accumulator),
                     "removal records must validate, i = {}",
                     i
                 );
-                assert!(accumulator.kernel.can_remove(removal_record));
+                assert!(accumulator.can_remove(removal_record));
             }
         }
 
@@ -624,10 +623,8 @@ mod removal_record_tests {
         assert!(original_first_removal_record
             .as_ref()
             .unwrap()
-            .validate(&accumulator.kernel));
-        assert!(!accumulator
-            .kernel
-            .can_remove(&original_first_removal_record.unwrap()));
+            .validate(&accumulator));
+        assert!(!accumulator.can_remove(&original_first_removal_record.unwrap()));
     }
 
     #[test]
